chore: Moved Map type definitions and interfaces into seperate type file
